mcp-server/requirements.txt
fastmcp==0.4.1
httpx==0.27.0
python-dotenv==1.0.1
